mode OD to IS module
