+++ library/src/otx/backend/native/models/detection/heads/__init__.py
@@ -6,7 +6,6 @@
 from .atss_head import ATSSHead
 from .dfine_decoder import DFINETransformer
 from .rtdetr_decoder import RTDETRTransformer
-from .rtmdet_head import RTMDetSepBNHead
 from .ssd_head import SSDHead
 from .yolox_head import YOLOXHead
 
@@ -14,7 +13,6 @@ __all__ = [
     "ATSSHead",
     "DFINETransformer",
     "RTDETRTransformer",
-    "RTMDetSepBNHead",
     "SSDHead",
     "YOLOXHead",
 ]
