+++ library/src/otx/backend/native/models/detection/necks/__init__.py
@@ -3,9 +3,8 @@
 
 """Custom neck implementations for detection task."""
 
-from .cspnext_pafpn import CSPNeXtPAFPN
 from .fpn import FPN
 from .hybrid_encoder import HybridEncoder
 from .yolox_pafpn import YOLOXPAFPN
 
-__all__ = ["FPN", "YOLOXPAFPN", "CSPNeXtPAFPN", "HybridEncoder"]
+__all__ = ["FPN", "YOLOXPAFPN", "HybridEncoder"]
