+++ library/src/otx/backend/native/models/instance_segmentation/necks/__init__.py
@@ -1,0 +1,8 @@
+# Copyright (C) 2026 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+"""Custom neck implementations for instance segmentation task."""
+
+from .cspnext_pafpn import CSPNeXtPAFPN
+
+__all__ = ["CSPNeXtPAFPN"]
