+++ library/src/otx/backend/native/models/instance_segmentation/rtmdet_inst.py
@@ -20,10 +20,10 @@ from otx.backend.native.models.common.utils.prior_generators import MlvlPointGen
 from otx.backend.native.models.common.utils.samplers import PseudoSampler
 from otx.backend.native.models.detection.backbones import CSPNeXt
 from otx.backend.native.models.detection.detectors import SingleStageDetector
-from otx.backend.native.models.detection.necks import CSPNeXtPAFPN
 from otx.backend.native.models.instance_segmentation.base import OTXInstanceSegModel
 from otx.backend.native.models.instance_segmentation.heads import RTMDetInstSepBNHead
 from otx.backend.native.models.instance_segmentation.losses import DiceLoss, RTMDetInstCriterion
+from otx.backend.native.models.instance_segmentation.necks import CSPNeXtPAFPN
 from otx.backend.native.models.modules.norm import build_norm_layer
 from otx.backend.native.models.utils.utils import load_checkpoint
 from otx.config.data import TileConfig
